+++ src/functions/index.ts
@@ -1,5 +1,6 @@
 import { ModuleRegistrationContext } from '@causa/workspace';
 import { EventTopicGenerateCodeForTypeScriptAndJsonEvents } from './event-topic-generate-json.js';
+import { ProjectBuildArtefactForServiceContainer } from './project-build-artefact-service-container.js';
 import { ProjectBuildArtefactForTypeScriptPackage } from './project-build-artefact-ts-package.js';
 import { ProjectGetArtefactDestinationForNpmPackage } from './project-get-artefact-destination-npm-package.js';
 import { ProjectPushArtefactForNpmPackage } from './project-push-artefact-npm-package.js';
@@ -8,6 +9,7 @@ import { ProjectReadVersionForJavascript } from './project-read-version-javascri
 export function registerFunctions(context: ModuleRegistrationContext) {
   context.registerFunctionImplementations(
     EventTopicGenerateCodeForTypeScriptAndJsonEvents,
+    ProjectBuildArtefactForServiceContainer,
     ProjectBuildArtefactForTypeScriptPackage,
     ProjectGetArtefactDestinationForNpmPackage,
     ProjectPushArtefactForNpmPackage,